% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkcluster.R
\docType{package}
\name{parkcluster}
\alias{parkcluster}
\alias{parkcluster-package}
\title{parkcluster: A permutation test for significance of clusters}
\description{
For information on the algorithm, read this paper:
}
\details{
Park, P.J., Manjourides, J., Bonetti, M., and Pagano, M.
A permutation test for determining significance of clusters with
applications to spatial and gene expression data.
Journal of Computational Statistics and Data Analysis. 53:4290-4300, 2009.
}
\section{parkcluster functions}{

phclust_pvalues - calculate p-values for each branch
phclust_plot - plot a tree showinga clusters and branch significance
}

