% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_code.R
\name{phclust_plot}
\alias{phclust_plot}
\title{Plot the hierarchical cluster significance tree}
\usage{
phclust_plot(x, pv = NULL, group = NULL, cutoff = 1, show.pval = TRUE,
  ...)
}
\arguments{
\item{x}{A data.frame, matrix, or hclust object}

\item{pv}{A vector of p-values, an output of phclust_pvalues}

\item{group}{A factor defining a hypothesized group, these groups will be colored on the tree, but will not affect the main algorithm}

\item{cutoff}{Minimum p-value required to resolved a branch}

\item{show.pval}{Whether to show the branch p-values on the tree}

\item{...}{Arguments that will be passed to phclust_pvalues (ignored if pv is given)}
}
\value{
A tree plot
}
\description{
Plot the hierarchical cluster significance tree
}
\examples{
phclust_plot(cars)
phclust_plot(iris[1:3], group=iris$Species, cutoff=0.1, ntrials=10)
}

