% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test_code.R
\name{phclust_pvalues}
\alias{phclust_pvalues}
\title{Calculate p-values for each branch in the tree}
\usage{
phclust_pvalues(x, hcl = phclust(x), method = "svd", ntrials = 100)
}
\arguments{
\item{x}{A data.frame or matrix with one row for each individual}

\item{hcl}{A "hclust" distance object, it will be created if missing.}

\item{method}{Either 'trace' or 'svd', see Park 2009}

\item{ntrials}{The number of simulations to use for predicting branch confidence}
}
\value{
A vector of p-values
}
\description{
Calculate p-values for each branch in the tree
}
\examples{
phclust_pvalues(iris[1:3], ntrials=10)
}

